Move $(embedExecutable "etcd") in dedicated module
HLS seems to load the whole executable when checking the template
haskell splice. This will make this only choke when loading this
Hydra.Network.EtcdBinary into HLS.
